DAS-2070: Migrate open_datatree into xarray.
First stab. Will need to add/move tests.
