+++ xarray/__init__.py
@@ -6,6 +6,7 @@ from xarray.backends.api import (
     load_dataset,
     open_dataarray,
     open_dataset,
+    open_datatree,
     open_mfdataset,
     save_mfdataset,
 )
@@ -82,6 +83,7 @@ __all__ = (
     "ones_like",
     "open_dataarray",
     "open_dataset",
+    "open_datatree",
     "open_mfdataset",
     "open_zarr",
     "polyval",
