+++ xarray/datatree_/datatree/__init__.py
@@ -1,9 +1,8 @@
 # import public API
 from .datatree import DataTree
 from .extensions import register_datatree_accessor
-from .io import open_datatree
 from .mapping import TreeIsomorphismError, map_over_subtree
-from .treenode import InvalidTreeError, NotFoundInTreeError
+from .treenode import InvalidTreeError, NodePath, NotFoundInTreeError
 
 try:
     # NOTE: the `_version.py` file must not be present in the git repository
@@ -15,10 +14,10 @@ except ImportError:  # pragma: no cover
 
 __all__ = (
     "DataTree",
-    "open_datatree",
     "TreeIsomorphismError",
     "InvalidTreeError",
     "NotFoundInTreeError",
+    "NodePath",
     "map_over_subtree",
     "register_datatree_accessor",
     "__version__",
