+++ xarray/datatree_/datatree/tests/test_io.py
@@ -1,7 +1,7 @@
 import pytest
 import zarr.errors
 
-from datatree.io import open_datatree
+from xarray import open_datatree
 from datatree.testing import assert_equal
 from datatree.tests import requires_h5netcdf, requires_netCDF4, requires_zarr
